Asynchronous drop storage table (#545)
1. A couple of virtual methods were introduced to make the behavior of dropping the storage table consistent between Cassandra and DynamoDB.
2. Since the logic of cleaning defunct storage tables is identical, the implementation was moved to the base class of the storage handler.
3. Make a slight change on TxWorkerPool. It is better to drain the task queue before quitting.

+++ include/table_statistics.h
@@ -75,7 +75,7 @@ public:
           units_(Units(param.records_)),
           sample_pool_(param.sample_keys_)
     {
-        assert(units_ >= sample_pool_.Size());
+        assert(units_ >= static_cast<int64_t>(sample_pool_.Size()));
     }
 
     void Reset(SamplePool &&sample_pool,
@@ -155,7 +155,7 @@ public:
         }
         units_ += Units(param.records_);
         sample_pool_.ClearCounter();
-        assert(units_ >= sample_pool_.Size());
+        assert(units_ >= static_cast<int64_t>(sample_pool_.Size()));
     }
 
     void Prune(const SamplePoolParam<KeyT> &param)
@@ -166,7 +166,7 @@ public:
         }
         units_ -= std::min(units_, static_cast<int64_t>(Units(param.records_)));
         sample_pool_.ClearCounter();
-        assert(units_ >= sample_pool_.Size());
+        assert(units_ >= static_cast<int64_t>(sample_pool_.Size()));
     }
 
     void To(remote::NodeGroupSamplePool *remote_ccmap_sample_pool) const
